requirements.txt
Pillow==9.3.0
python_telegram_bot==13.14
requests==2.22.0
telegram==0.0.1
